# openapi/heartimate.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Heartimate API",
    "version": "0.1.0",
    "description": "Authenticated JSON APIs plus a streaming chat endpoint. All JSON responses include `success`. Errors include `error`. Responses include `x-request-id` for correlation."
  },
  "servers": [{ "url": "/", "description": "App origin" }],
...
  "components": {
    "securitySchemes": {
      "clerkSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Clerk session cookie (browser) or Bearer from Clerk"
      }
    },
...
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "const": false },
          "error": { "type": "string" },
          "code": {
            "type": "string",
            "description": "Machine-readable code (see lib/api-error-codes.ts)"
          },
          "retryAfter": { "type": "integer", "description": "Seconds (429 only)" }
        }
      },
...
      "OnboardingPostRequest": {
        "type": "object",
        "required": ["displayName", "personaId", "starterCharId", "isAdult"],
        "properties": {
          "displayName": { "type": "string", "maxLength": 50 },
          "kinkPrefs": { "type": "array", "items": { "type": "string" } },
          "personaId": { "type": "string" },
          "starterCharId": { "type": "string" },
          "characterName": { "type": "string", "maxLength": 80 },
          "isAdult": { "type": "boolean", "const": true }
        }
      },
...
    },
    "responses": {
...
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" } }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
        }
      }
    }
  },
  "security": [{ "clerkSession": [] }],
  "paths": {
...
    "/api/onboarding": {
      "get": {
        "tags": ["onboarding"],
        "security": [],
        "responses": {
          "200": { "description": "Public starter characters" }
        }
      },
      "post": {
        "tags": ["onboarding"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OnboardingPostRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "const": true },
                    "chatId": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/characters": {
      "post": {
        "tags": ["characters"],
        "summary": "Create character (multipart)",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "avatar": { "type": "string", "format": "binary" },
                  "isPublic": { "type": "string" },
                  "forkedFromId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created; body includes character id" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  }
}